+++ MarkEquipsAPI/MarkEquipsAPI/Services/IEntitieService.cs
@@ -6,10 +6,10 @@ namespace MarkEquipsAPI.Services
     public interface IEntitieService
     {
         Equipment Create(Equipment equipment);
-        Equipment FindByID(long id);
+        Equipment FindByID(int id);
         List<Equipment> FindAll();
         Equipment Update(Equipment equipment);
-        void Delete(long id);
+        void Delete(int id);
 
     }
 }
